Add hdf5 convert option

+++ convert.py
@@ -9,6 +9,7 @@ import caffe
 from caffe.proto import caffe_pb2
 import numpy as np
 import os, sys, getopt
+import h5py
 
 # Check dataset information before checking
 def lmCheck(db_name):
@@ -51,32 +52,47 @@ def levCheck(db_name):
     print 'Total features: ', count
 
 # Convert dataset
-def levSave2txt(db_name, outputfile):
+def levSave2txt(db_name, outputfile, file_type):
     db = leveldb.LevelDB(db_name)
     count = 0
-    with open(outputfile, 'w') as writer:
-        writer.truncate()
-        for key, val in db.RangeIter():
-            #count = count + 1
-            #print count 
-            datum = caffe.io.caffe_pb2.Datum() 
-            datum.ParseFromString(val) 
-            data = caffe.io.datum_to_array(datum)
-            np.savetxt(writer, data.reshape(1,-1), fmt='%.8g')
-        #print count
-
+    
+    if file_type == 'text':
+        with open(outputfile, 'w') as writer:
+            writer.truncate()
+            for key, val in db.RangeIter():
+                #count = count + 1
+                #print count 
+                datum = caffe.io.caffe_pb2.Datum() 
+                datum.ParseFromString(val) 
+                data = caffe.io.datum_to_array(datum)
+                np.savetxt(writer, data.reshape(1,-1), fmt='%.8g')
+            #print count
+    elif file_type == 'hdf5':
+        with h5py.File(outputfile, 'w') as hf:
+            data_matrix = []
+            for key, val in db.RangeIter():
+                #count = count + 1
+                #print count 
+                datum = caffe.io.caffe_pb2.Datum() 
+                datum.ParseFromString(val) 
+                data = caffe.io.datum_to_array(datum)
+                data_matrix.append(data)
+            data_matrix = np.array(data_matrix)
+            hf.create_dataset('dataset',data=data_matrix)
+            #print count
 def main(argv):
     inputfile = ''
     outputfile = ''
 
     try:
-        opts, args = getopt.getopt(argv,"hi:o:m:t:",["ifile=","ofile="])
+        opts, args = getopt.getopt(argv,"hi:o:m:t:w:",["ifile=","ofile="])
     except getopt.GetoptError:
         print '======================================================================================='
         print 'Usage:'
-        print 'caffe_feature_extractor.py -i <inputfile> -o <outputfile> -t <dataset type> -m <mode>'
+        print 'convert.py -i <inputfile> -o <outputfile> -t <dataset type> -m <mode> -w <file type>'
         print 'mode -- check, convert'
         print 'dataset type -- leveldb, lmdb'
+        print 'file type -- text, hdf5'
         print '======================================================================================='
         sys.exit(2)
 
@@ -85,9 +101,10 @@ def main(argv):
         if opt == '-h':
             print '======================================================================================='
             print 'Usage:'
-            print 'caffe_feature_extractor.py -i <inputfile> -o <outputfile> -t <dataset type> -m <mode>'
+            print 'convert.py -i <inputfile> -o <outputfile> -t <dataset type> -m <mode> -w <file type>'
             print 'mode -- check, convert'
             print 'dataset type -- leveldb, lmdb'
+            print 'file type -- text, hdf5'
             print '======================================================================================='
             sys.exit()
         elif opt in ("-i"):
@@ -102,13 +119,16 @@ def main(argv):
         elif opt in ("-m"):
             mode = arg
             print 'Function mode is', mode
+        elif opt in ("-w"):
+            file_type = arg
+            print 'Function mode is', file_type
     print '\n'
 
     if str.upper(dataset_type) == "LEVELDB":
         if str.upper(mode) == "CHECK": 
             levCheck(inputfile)
         elif str.upper(mode) == "CONVERT":
-            levSave2txt(inputfile, outputfile)
+            levSave2txt(inputfile, outputfile, file_type)
 
     elif str.upper(dataset_type) == "LMDB":
         if str.upper(mode) == "CHECK":
